# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: HireKit API
  version: 1.0.0
  description: |
    The REST API the HireKit web app calls: sign-in, roles and criteria, resume upload,
    the ranked candidate list, recruiter decisions, the interview kit, interviewer feedback,
    comparison, jobs and the model cost log. Two roles use it: the recruiter and the
    interviewer. An interviewer never receives resume text, evidence quotes, notes, audit
    history, file names or another interviewer's feedback.
    Serves: US-00-001 to US-00-016, US-02-002 (AC-US-02-002-4). Design: docs/design/api-lld.md.
    Style: docs/api/README.md.
  contact:
    name: HireKit maintainers
  x-api-style: REST over HTTPS with JSON bodies, one service, one React client generated from this file (ADR-0006).
  x-versioning: >-
    Major in the path (/v1). No dated behaviour versions in the first build; the first breaking
    change is planned with the api-versioning skill.
  x-authentication: >-
    A server-side session in the HttpOnly cookie hirekit_session (ADR-0005). Every POST, PUT and
    DELETE also sends the session's csrf_token in the X-CSRF-Token header. A record the caller may
    not see is 404, never 403, so existence is not leaked.
  x-conventions:
    - rule: No Idempotency-Key header on creating POSTs.
...
    - rule: The ranked list and the job queue view use limit and offset, not cursors.
...
    - rule: No X-API-Version and no rate-limit headers in the first build.
...
servers:
  - url: http://localhost:8080
...
security:
  - cookieAuth: []
    csrfToken: []
...
paths:
...
  /v1/roles/{role_id}/candidates:
    get:
      operationId: listCandidates
      summary: The ranked list
      description: >-
        Every candidate of the role, ranked by weighted total with overrides applied; low scorers
        stay visible. Scores from an older criteria version are returned and marked stale. Offset
        pagination is used because the order is computed; limit is at most 200 and offset at most
        1,000.
      tags: [candidates]
      x-story-ids: [US-00-008, US-00-009]
      x-roles: [recruiter]
      parameters:
        - $ref: '#/components/parameters/RoleId'
        - name: filter[stage]
          in: query
          schema: { $ref: '#/components/schemas/Stage' }
        - name: sort
          in: query
          description: total (default, descending) or a criterion id
          schema: { type: string, default: total }
        - $ref: '#/components/parameters/Limit200'
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, maximum: 1000, default: 0 }
      responses:
        '200':
          description: One page of the ranked list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RankedList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/me/candidates:
    get:
      operationId: listMyCandidates
      summary: The interviewer's assigned candidates (at most 200)
      tags: [candidates]
      x-story-ids: [US-00-012, US-00-016]
      x-roles: [interviewer]
      parameters:
        - $ref: '#/components/parameters/Limit200'
      responses:
        '200':
          description: The candidates assigned to the caller
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MyCandidates' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/candidates/{candidate_id}:
    parameters:
      - $ref: '#/components/parameters/CandidateId'
    get:
      operationId: getCandidate
      summary: One candidate
      description: >-
        A recruiter gets every field. An interviewer gets it only for a candidate assigned to them,
        and never receives quotes, notes, flag reasons, audit history, file names, resume text, or
        (until they submit feedback) any model score.
      tags: [candidates]
      x-story-ids: [US-00-008, US-00-009, US-00-012, US-00-014]
      x-roles: [recruiter, interviewer]
      responses:
        '200':
          description: The candidate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CandidateDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /v1/candidates/{candidate_id}/scores/{criterion_id}/override:
    put:
      operationId: overrideScore
      summary: Override a criterion score with a note
      tags: [candidates]
      x-story-ids: [US-00-010]
      x-roles: [recruiter]
      parameters:
        - $ref: '#/components/parameters/CandidateId'
        - $ref: '#/components/parameters/CriterionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OverrideRequest' }
      responses:
        '200':
          description: The updated score cell
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScoreCell' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /v1/candidates/{candidate_id}/stage:
    post:
      operationId: changeStage
      summary: Change the stage (the only route that writes it)
      tags: [candidates]
      x-story-ids: [US-00-011]
      x-roles: [recruiter]
      parameters:
        - $ref: '#/components/parameters/CandidateId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StageRequest' }
      responses:
        '200':
          description: The new stage
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StageChanged' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /v1/candidates/{candidate_id}:reveal-identity:
    post:
      operationId: revealIdentity
      summary: Reveal the candidate's name (audited before it is returned)
      tags: [candidates]
      x-story-ids: [US-00-009]
      x-roles: [recruiter]
      parameters:
        - $ref: '#/components/parameters/CandidateId'
      responses:
        '200':
          description: The identity (personal data)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Identity' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /v1/candidates/{candidate_id}/assignments:
    post:
      operationId: assignInterviewer
      summary: Assign an interviewer (idempotent)
      description: >-
        The target must be an interviewer, else 422 validation_error. A repeat returns 201 again
        and changes nothing. An unknown candidate is 404.
      tags: [candidates]
      x-story-ids: [US-00-016]
      x-roles: [recruiter]
      parameters:
        - $ref: '#/components/parameters/CandidateId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignRequest' }
      responses:
        '201':
          description: Assigned
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Assignment' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /v1/candidates/{candidate_id}/assignments/{user_id}:
    delete:
      operationId: removeAssignment
      summary: Remove an assignment
      description: >-
        The candidate leaves that interviewer's list at once. Removing an absent pair is 204; an
        unknown candidate is 404.
      tags: [candidates]
      x-story-ids: [US-00-016]
      x-roles: [recruiter]
      parameters:
        - $ref: '#/components/parameters/CandidateId'
        - name: user_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Removed }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: hirekit_session
      description: Server-side session cookie (HttpOnly, SameSite=Lax; Secure outside local development).
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: The csrf_token from login or GET /v1/auth/me; required on every POST, PUT and DELETE.
  parameters:
    RoleId: { name: role_id, in: path, required: true, schema: { type: string, format: uuid } }
    CandidateId: { name: candidate_id, in: path, required: true, schema: { type: string, format: uuid } }
    CriterionId: { name: criterion_id, in: path, required: true, schema: { type: string, format: uuid } }
...
    Limit200:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 100 }
  responses:
    Unauthorized:
      description: No signed-in user, or the session expired
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { error: { code: unauthenticated, message: sign in required, details: {}, request_id: req_01J9Z6 } }
    Forbidden:
      description: Signed in but not allowed (wrong role, or a missing or wrong CSRF token)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { error: { code: forbidden, message: this role cannot do that, details: {}, request_id: req_01J9Z6 } }
    NotFound:
      description: Not found, or not visible to this caller (existence is never revealed)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { error: { code: not_found, message: not found, details: {}, request_id: req_01J9Z6 } }
    Conflict:
      description: The request conflicts with the current state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { error: { code: role_not_approved, message: approve the criteria first, details: {}, request_id: req_01J9Z6 } }
...
    Unprocessable:
      description: Valid shape but invalid content
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          example: { error: { code: validation_error, message: request failed validation, details: { errors: [] }, request_id: req_01J9Z6 } }
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: true }
            request_id: { type: [string, 'null'] }
      example: { error: { code: not_found, message: not found, details: {}, request_id: req_01J9Z6 } }
      x-error-codes:
        - { code: unauthenticated, status: 401, meaning: "No signed-in user, an expired session, or a wrong email or password" }
        - { code: csrf_failed, status: 403, meaning: "The X-CSRF-Token header is missing or wrong" }
        - { code: forbidden, status: 403, meaning: "The caller's role may not use this route" }
        - { code: not_found, status: 404, meaning: "Missing, or not visible to this caller" }
        - { code: role_not_approved, status: 409, meaning: "The role is Draft; approve the criteria first" }
        - { code: scores_stale, status: 409, meaning: "The score belongs to an older criteria version" }
        - { code: same_stage, status: 409, meaning: "The candidate is already in that stage" }
        - { code: job_already_open, status: 409, meaning: "A scoring, proposal or kit job is already open" }
        - { code: not_retryable, status: 409, meaning: "The candidate does not need scoring" }
        - { code: job_not_cancellable, status: 409, meaning: "The job is already finished" }
        - { code: feedback_locked, status: 409, meaning: "Feedback is already submitted and locked" }
        - { code: criteria_changed, status: 409, meaning: "The criteria changed since the caller loaded them; details.current_version has the current one" }
        - { code: no_criteria, status: 422, meaning: "The role has no live criteria to approve" }
        - { code: incomplete_rubric, status: 422, meaning: "A live criterion lacks a descriptor for a level 0 to 4; details.criterion_ids lists them" }
        - { code: budget_reached, status: 409, meaning: "The model budget is reached in live mode" }
        - { code: incomplete_feedback, status: 422, meaning: "A criterion has no score or comment" }
        - { code: too_many_files, status: 422, meaning: "More than 100 files in one upload" }
        - { code: no_criteria, status: 422, meaning: "The role has no criteria to approve" }
        - { code: validation_error, status: 422, meaning: "A field failed validation" }
        - { code: payload_too_large, status: 413, meaning: "The request body is over the cap" }
        - { code: internal, status: 500, meaning: "An unexpected error; quote the request_id" }
        - { code: service_unavailable, status: 503, meaning: "The database cannot be reached" }
...
    Stage:
      type: string
      enum: [new, screened, interview, offer, hired, rejected, withdrawn]
      example: screened
...
    ScoreCell:
      type: object
      required: [criterion_id, criterion_name, kind, status, source]
      properties:
        criterion_id: { type: string, format: uuid }
        criterion_name: { type: string }
        kind: { type: string, enum: [must_have, nice_to_have] }
        status: { type: string, enum: [scored, no_evidence, failed] }
        model_score: { type: [integer, 'null'], minimum: 0, maximum: 4 }
        override_score: { type: [integer, 'null'], minimum: 0, maximum: 4 }
        source: { type: string, enum: [model_suggestion, no_evidence_found, recruiter_override, failed] }
        stale: { type: boolean }
        quote: { type: [string, 'null'], description: Recruiter only; the verified evidence quote }
        flag_reason: { type: [string, 'null'], description: Recruiter only }
        override_note: { type: [string, 'null'], description: Recruiter only }
      example:
        criterion_id: 20000000-0000-4000-8000-000000000001
        criterion_name: Python experience
        kind: must_have
        status: scored
        model_score: 3
        override_score: null
        source: model_suggestion
        stale: false
        quote: Built a payments service in Python at Acme.
        flag_reason: null
        override_note: null
    RankedCandidate:
      type: object
      required: [id, candidate_no, stage, processing_status, total, must_have_covered, must_have_total, stale, scores]
      properties:
        id: { type: string, format: uuid }
        candidate_no: { type: integer }
        stage: { $ref: '#/components/schemas/Stage' }
        processing_status: { type: string, enum: [queued, parsing, anonymizing, scoring, done, failed] }
        failure_reason: { type: [string, 'null'] }
        total: { type: number }
        must_have_covered: { type: integer }
        must_have_total: { type: integer }
        stale: { type: boolean, description: True when the scores are from an older criteria version }
        duplicate_of_candidate_no: { type: [integer, 'null'] }
        scores:
          type: array
          items: { $ref: '#/components/schemas/ScoreCell' }
      example:
        id: 30000000-0000-4000-8000-000000000001
        candidate_no: 14
        stage: screened
        processing_status: done
        failure_reason: null
        total: 11.5
        must_have_covered: 2
        must_have_total: 3
        stale: false
        duplicate_of_candidate_no: null
        scores: []
    RankedList:
      type: object
      required: [data, page]
      properties:
        data: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/RankedCandidate' } }
        page:
          type: object
          required: [limit, offset, total]
          properties:
            limit: { type: integer }
            offset: { type: integer }
            total: { type: integer }
      example:
        data: []
        page: { limit: 100, offset: 0, total: 0 }
    MyCandidates:
      type: object
      required: [data]
      properties:
        data:
          type: array
          maxItems: 200
          items:
            type: object
            required: [candidate_id, candidate_no, role_id, role_title, has_submitted]
            properties:
              candidate_id: { type: string, format: uuid }
              candidate_no: { type: integer }
              role_id: { type: string, format: uuid }
              role_title: { type: string }
              has_submitted: { type: boolean }
      example:
        data:
          - { candidate_id: 30000000-0000-4000-8000-000000000001, candidate_no: 14, role_id: 10000000-0000-4000-8000-000000000001, role_title: Backend engineer, has_submitted: false }
    AuditEvent:
      type: object
      required: [id, kind, actor_id, created_at]
      properties:
        id: { type: integer, format: int64 }
        kind: { type: string, enum: [score_override, stage_change, identity_reveal, feedback_edit_approved, feedback_edited] }
        actor_id: { type: string, format: uuid }
        criterion_name: { type: [string, 'null'] }
        old_score: { type: [integer, 'null'] }
        new_score: { type: [integer, 'null'] }
        from_stage: { oneOf: [{ $ref: '#/components/schemas/Stage' }, { type: 'null' }] }
        to_stage: { oneOf: [{ $ref: '#/components/schemas/Stage' }, { type: 'null' }] }
        note: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
      example:
        id: 12
        kind: stage_change
        actor_id: 00000000-0000-4000-8000-000000000001
        criterion_name: null
        old_score: null
        new_score: null
        from_stage: new
        to_stage: screened
        note: null
        created_at: 2026-09-30T10:00:00Z
    CandidateDetail:
      type: object
      description: >-
        A recruiter gets every field. An interviewer gets only id, candidate_no, role_id and
        has_submitted until they submit feedback, then scores too; stage, processing_status and
        audit are left out for them, not sent as null.
      required: [id, candidate_no, role_id]
      properties:
        id: { type: string, format: uuid }
        candidate_no: { type: integer }
        role_id: { type: string, format: uuid }
        stage: { $ref: '#/components/schemas/Stage' }
        processing_status: { type: string, enum: [queued, parsing, anonymizing, scoring, done, failed] }
        scores:
          type: array
          description: For an interviewer, present only after they submit feedback, and without quote, flag reason or note
          items: { $ref: '#/components/schemas/ScoreCell' }
        audit: { type: array, description: 'Recruiter only', items: { $ref: '#/components/schemas/AuditEvent' } }
        has_submitted: { type: boolean, description: Interviewer view }
      example:
        id: 30000000-0000-4000-8000-000000000001
        candidate_no: 14
        role_id: 10000000-0000-4000-8000-000000000001
        stage: screened
        processing_status: done
        scores: []
        audit: []
        has_submitted: false
...
    OverrideRequest:
      type: object
      required: [override_score, note]
      additionalProperties: false
      properties:
        override_score: { type: integer, minimum: 0, maximum: 4 }
        note: { type: string, minLength: 10 }
      example: { override_score: 4, note: 'Led the migration, seen in the interview notes' }
    StageRequest:
      type: object
      required: [stage]
      additionalProperties: false
      properties:
        stage: { $ref: '#/components/schemas/Stage' }
        reason: { type: [string, 'null'], description: Optional; recorded in the audit history }
      example: { stage: rejected, reason: Missing a must-have }
    StageChanged:
      type: object
      required: [candidate_id, from_stage, to_stage]
      properties:
        candidate_id: { type: string, format: uuid }
        from_stage: { $ref: '#/components/schemas/Stage' }
        to_stage: { $ref: '#/components/schemas/Stage' }
      example: { candidate_id: 30000000-0000-4000-8000-000000000001, from_stage: new, to_stage: screened }
    Identity:
      type: object
      required: [identity_name, file_name]
      properties:
        identity_name: { type: [string, 'null'] }
        file_name: { type: string }
      example: { identity_name: Jane Doe, file_name: Jane_Doe_CV.pdf }
    AssignRequest:
      type: object
      required: [user_id]
      additionalProperties: false
      properties:
        user_id: { type: string, format: uuid, description: Must be an interviewer }
      example: { user_id: 00000000-0000-4000-8000-000000000002 }
    Assignment:
      type: object
      required: [candidate_id, user_id]
      properties:
        candidate_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
      example: { candidate_id: 30000000-0000-4000-8000-000000000001, user_id: 00000000-0000-4000-8000-000000000002 }
...